+++ src/index.js
@@ -1,13 +1,13 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App';
-import Volunteer from './Volunteer'
-import Donation from './Donation'
-import Projects from './Projects'
+import VolunteerPage from './Pages/VolunteerPage'
+import DonationPage from './Pages/DonationPage'
+import ProjectsPage from './Pages/ProjectsPage'
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <React.StrictMode>
-    <Donation />
+    <DonationPage />
   </React.StrictMode>
 );
